# api/openapi.yaml
openapi: 3.1.0
info:
  title: KubeVirt Shepherd API
  description: |
    API for managing KubeVirt virtual machines and cloud-native governance.
    Adheres to ADR-0021: API Contract-First Development.
  version: 1.0.0
  contact:
    name: KubeVirt Shepherd Team
    url: https://github.com/kv-shepherd/shepherd
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api/v1
    description: Production API
...
paths:
...
  /admin/auth-providers/{provider_id}/directory/descriptor:
    get:
      tags: [auth-providers, admin]
      summary: Get directory sync capability descriptor for an auth provider
      operationId: getAuthProviderDirectoryDescriptor
      parameters:
        - $ref: '#/components/parameters/ProviderID'
      responses:
        '200':
          description: Directory sync capability descriptor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorySyncDescriptor'
        '404':
          $ref: '#/components/responses/NotFound'
        '501':
          description: Auth provider does not implement directory sync
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/auth-providers/{provider_id}/directory/schedule:
    get:
      tags: [auth-providers, admin]
      summary: Get scheduled directory enrichment status for an auth provider
      operationId: getAuthProviderDirectorySchedule
      parameters:
        - $ref: '#/components/parameters/ProviderID'
      responses:
        '200':
          description: Scheduled enrichment plan and runtime status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectoryEnrichmentScheduleStatus'
        '404':
          $ref: '#/components/responses/NotFound'
        '501':
          description: Auth provider does not implement directory sync
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/auth-providers/{provider_id}/directory/preview:
    post:
      tags: [auth-providers, admin]
      summary: Preview canonical directory import result
      operationId: previewAuthProviderDirectory
      parameters:
        - $ref: '#/components/parameters/ProviderID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectorySyncPreviewRequest'
      responses:
        '200':
          description: Canonical preview result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorySyncPreview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '501':
          description: Auth provider does not implement directory sync preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/auth-providers/{provider_id}/directory/sync:
    post:
      tags: [auth-providers, admin]
      summary: Trigger asynchronous directory sync job
      operationId: syncAuthProviderDirectory
      parameters:
        - $ref: '#/components/parameters/ProviderID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectorySyncRequest'
      responses:
        '202':
          description: Directory sync job enqueued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorySyncStartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '501':
          description: Auth provider does not implement directory sync
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/auth-providers/{provider_id}/directory/sync-jobs:
    get:
      tags: [auth-providers, admin]
      summary: List directory sync jobs for an auth provider
      operationId: listAuthProviderDirectorySyncJobs
      parameters:
        - $ref: '#/components/parameters/ProviderID'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Directory sync job list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorySyncJobList'
        '404':
          $ref: '#/components/responses/NotFound'
        '501':
          description: Auth provider does not implement directory sync
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/auth-providers/{provider_id}/directory/sync-jobs/{job_id}:
    get:
      tags: [auth-providers, admin]
      summary: Get one directory sync job summary
      operationId: getAuthProviderDirectorySyncJob
      parameters:
        - $ref: '#/components/parameters/ProviderID'
        - $ref: '#/components/parameters/DirectorySyncJobID'
      responses:
        '200':
          description: Directory sync job summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorySyncJobDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '501':
          description: Auth provider does not implement directory sync
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable Parameters (ADR-0023 Pagination) ──────
  parameters:
    Page:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ProviderID:
      name: provider_id
      in: path
      required: true
      schema:
        type: string
...
    DirectorySyncJobID:
      name: job_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: Machine-readable error code (frontend handles i18n)
        message:
          type: string
          description: English log message (not for UI display)
        params:
          type: object
          additionalProperties: true
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    # ── System ──────────────────────────────────────
...
    DirectorySyncDescriptor:
      type: object
      required: [display_name, supports_preview]
      properties:
        display_name:
          type: string
        description:
          type: string
        supports_preview:
          type: boolean
        request_schema:
          type: object
          additionalProperties: true
    DirectoryEnrichmentScheduleStatus:
      type: object
      required: [supported, enabled]
      properties:
        supported:
          type: boolean
        enabled:
          type: boolean
        mode:
          type: string
          enum: [enrich_existing_only]
        join_key_type:
          type: string
          enum: [username]
        schedule_cron:
          type: string
        schedule_timezone:
          type: string
        next_run_at:
          type: string
          format: date-time
        last_job_id:
          type: string
        last_job_status:
          type: string
          enum: [pending, running, completed, failed]
        last_job_created_at:
          type: string
          format: date-time
        last_job_completed_at:
          type: string
          format: date-time
        pending_job_id:
          type: string
        pending_job_status:
          type: string
          enum: [pending, running]
    DirectorySyncConflict:
      type: object
      required: [code]
      properties:
        code:
          type: string
          enum: [same_external_identity, username_conflict, email_conflict, ambiguous_existing_user]
        field:
          type: string
        existing_user_id:
          type: string
        message:
          type: string
    DirectoryUserRecord:
      type: object
      required: [external_id, username, display_name]
      properties:
        external_id:
          type: string
        username:
          type: string
        display_name:
          type: string
        email:
          type: string
        cohorts:
          type: array
          items:
            $ref: '#/components/schemas/DirectoryExternalCohort'
        attributes:
          type: object
          additionalProperties: true
    DirectoryExternalCohort:
      type: object
      required: [kind, key]
      properties:
        kind:
          type: string
        key:
          type: string
        display_name:
          type: string
    DirectorySyncPreviewItem:
      type: object
      required: [record, match]
      properties:
        record:
          $ref: '#/components/schemas/DirectoryUserRecord'
        match:
          $ref: '#/components/schemas/DirectorySyncPreviewMatch'
        conflicts:
          type: array
          items:
            $ref: '#/components/schemas/DirectorySyncConflict'
        warnings:
          type: array
          items:
            type: string
    DirectorySyncPreviewMatch:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [create, update, blocked]
        existing_user_id:
          type: string
        matched_by:
          type: string
          enum: [external_id]
    DirectorySyncPreview:
      type: object
      required: [total_count, items]
      properties:
        total_count:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/DirectorySyncPreviewItem'
    DirectorySyncPreviewRequest:
      type: object
      required: [provider_request]
      properties:
        provider_request:
          type: object
          additionalProperties: true
        conflict_resolution:
          type: string
          default: skip
          enum: [skip]
    DirectorySyncRequest:
      type: object
      required: [provider_request]
      properties:
        provider_request:
          type: object
          additionalProperties: true
        conflict_resolution:
          type: string
          default: skip
          enum: [skip]
    DirectorySyncStartResponse:
      type: object
      required: [job_id, status]
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
    DirectoryActionSummary:
      type: object
      required: [create_count, update_count, blocked_count]
      properties:
        create_count:
          type: integer
        update_count:
          type: integer
        blocked_count:
          type: integer
    DirectorySyncJob:
      type: object
      required:
        [
          id,
          provider_id,
          status,
          conflict_resolution,
          total_entries,
          result_summary,
          error_count,
          errors,
          triggered_by,
          created_at,
          updated_at
        ]
      properties:
        id:
          type: string
        provider_id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        conflict_resolution:
          type: string
          enum: [skip]
        sync_mode:
          type: string
          enum: [manual_import, scheduled_enrichment]
        join_key_type:
          type: string
        total_entries:
          type: integer
        result_summary:
          $ref: '#/components/schemas/DirectoryActionSummary'
        error_count:
          type: integer
        errors:
          type: array
          items:
            type: string
        triggered_by:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DirectorySyncJobDetail:
      allOf:
        - $ref: '#/components/schemas/DirectorySyncJob'
        - type: object
          required: [request_snapshot]
          properties:
            request_snapshot:
              type: object
              additionalProperties: true
    DirectorySyncJobList:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DirectorySyncJob'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
security:
  - BearerAuth: []